+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Raw HTML support (#191)
 - Introduce new "uplink tree" view (#195)
 - Resilient error handling (#202)
+- Added `neuron query --graph` to get the entire graph as JSON export
 - Bug fixes
   - Fix 'neuron new' generating invalid Markdown when title contains special characters (#163)